+++ src/components/calendar/types.ts
@@ -1,0 +1,51 @@
+import type { Doc, Id } from "../../../convex/_generated/dataModel";
+import type { SchedulingSimulatedContext } from "../../types";
+
+export interface Appointment {
+  color: string;
+  column: string; // Resource ID (practitioner ID or "ekg" / "labor")
+  convexId?: Id<"appointments">; // Original Convex ID for real appointments
+  duration: number; // in minutes
+  id: string;
+  isSimulation: boolean;
+  replacesAppointmentId?: Id<"appointments"> | null;
+  resource?: {
+    appointmentType?: Doc<"appointments">["appointmentType"];
+    isSimulation?: boolean;
+    locationId?: Doc<"appointments">["locationId"];
+    patientId?: Doc<"appointments">["patientId"];
+    practitionerId?: Doc<"appointments">["practitionerId"];
+  };
+  startTime: string;
+  title: string;
+}
+
+export interface NewCalendarProps {
+  locationSlug?: string | undefined;
+  onDateChange?: (date: Date) => void;
+  onLocationResolved?: (
+    locationId: Id<"locations">,
+    locationName: string,
+  ) => void;
+  onUpdateSimulatedContext?: (context: SchedulingSimulatedContext) => void;
+  practiceId?: Id<"practices">;
+  selectedLocationId?: Id<"locations"> | undefined;
+  showGdtAlert?: boolean;
+  simulatedContext?: SchedulingSimulatedContext;
+  simulationDate?: Date;
+}
+
+export const SLOT_DURATION = 5; // minutes
+
+export const APPOINTMENT_COLORS = [
+  "bg-blue-500",
+  "bg-green-500",
+  "bg-purple-500",
+  "bg-orange-500",
+  "bg-pink-500",
+  "bg-teal-500",
+  "bg-red-500",
+  "bg-yellow-500",
+  "bg-indigo-500",
+  "bg-gray-500",
+];
